+++ test-helpers/sample.cjs
@@ -1,9 +1,8 @@
-/* eslint-disable import/unambiguous -- Testing */
 'use strict';
 
 const a = 0;
 
-/* c8 ignore next 3 */
+/* c8 ignore next 3 -- Some description */
 if (a) {
   console.log('A branch test');
 } else {